requirements.txt
configargparse
mss
numpy
pyserial
Pillow
pygame
tensorflow==1.13.1
